add pattern mvvm
